Issue #3891: Split and Organize Checkstyle inputs by Test: naming package, TypeName
