+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/naming/typename/InputTypeName.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.naming;
+package com.puppycrawl.tools.checkstyle.checks.naming.typename;
 
 class inputHeaderClass {
